[accel-wave] split worker.ts
