+++ packages/accel-wave/src/index.ts
@@ -1,4 +1,4 @@
 export { Config } from "./config.js";
-export { worker } from "./worker.js";
+export { worker } from "./worker/index.js";
 export { mount } from "./mount.js";
 export { BaseUploader, Item } from "./uploader.js";
